+++ CHANGELOG.md
@@ -2,4 +2,7 @@
 
 **v1.0.0**
 
-- Basic state machine system with goroutines
+- Basic state machine system
+- Supports async action
+- Supports sync action
+- Supports multi-reducers